Fixed: Set nohidden.
The "hidden" option causes Vim to keep a buffer, once open e.g. through
a split, open, even if the window that displayed the buffer is closed.
That causes swap file conflicts, if you try to open the same file in a
parallel Vim session. Setting "nohidden" avoids this issue.